LPS-53617 Move provider logic out from asset
This API can be used by portlets that do not use the asset framework.
