+++ modules/apps/journal/journal-content-web/src/com/liferay/journal/content/web/portlet/asset/provider/JournalContentAddPortletProvider.java
@@ -15,11 +15,11 @@
 package com.liferay.journal.content.web.portlet.asset.provider;
 
 import com.liferay.journal.content.web.constants.JournalContentPortletKeys;
+import com.liferay.portal.kernel.provider.AddPortletProvider;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portlet.asset.AssetRendererFactoryRegistryUtil;
 import com.liferay.portlet.asset.model.AssetEntry;
-import com.liferay.portlet.asset.provider.AddPortletProvider;
 import com.liferay.portlet.asset.service.AssetEntryLocalService;
 import com.liferay.portlet.journal.asset.JournalArticleAssetRenderer;
 import com.liferay.portlet.journal.asset.JournalArticleAssetRendererFactory;
