+++ modules/apps/journal/journal-content-web/src/com/liferay/journal/content/web/portlet/asset/provider/JournalContentDisplayInformationProvider.java
@@ -15,8 +15,8 @@
 package com.liferay.journal.content.web.portlet.asset.provider;
 
 import com.liferay.journal.content.web.constants.JournalContentPortletKeys;
+import com.liferay.portal.kernel.provider.DisplayInformationProvider;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portlet.asset.provider.DisplayInformationProvider;
 import com.liferay.portlet.journal.model.JournalArticle;
 
 import javax.portlet.PortletPreferences;
