+++ portal-service/src/com/liferay/portal/kernel/provider/packageinfo
